wired up sort url queries.
