+++ hooks/interfaces.ts
@@ -1,6 +1,8 @@
 type TClearBy = () => {}
 
 export interface IfilterCategories {
+  sortTab: () => void
+  sortTableRow: () => void
   homeType: () => void
   status: () => void
   beds: () => void
